+++ example/lib/main.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:hybrid_map_marker/hybrid_map_marker.dart';
+import 'package:hybrid_map_marker/hybrid_map_marker_impl.dart';
 
 void main() {
   runApp(const MainApp());
@@ -15,7 +16,7 @@ class MainApp extends StatefulWidget {
 }
 
 class _MainAppState extends State<MainApp> {
-  final HybridMapMarker _hybridMapMarker = HybridMapMarker.instance;
+  final HybridMapMarker _hybridMapMarker = HybridMapMarkerImpl.instance;
 
   @override
   Widget build(BuildContext context) {
@@ -71,8 +72,91 @@ class _MainAppState extends State<MainApp> {
     );
   }
 
+  Widget _jpgImageMarker() {
+    return Container(
+      margin: const EdgeInsets.all(2),
+      padding: EdgeInsets.all(8),
+      decoration: BoxDecoration(
+        color: Colors.amber,
+        shape: .circle,
+      ),
+      child: Container(
+        width: 64,
+        height: 64,
+        clipBehavior: Clip.antiAlias,
+        decoration: BoxDecoration(
+          color: Colors.amber,
+          shape: .circle,
+        ),
+        child: Image.asset(
+          'assets/bird.jpg',
+          fit: BoxFit.cover,
+        ),
+      ),
+    );
+  }
+
+  Widget _pngImageMarker() {
+    return Container(
+      margin: const EdgeInsets.all(2),
+      padding: EdgeInsets.all(8),
+      decoration: BoxDecoration(
+        color: Colors.amber,
+        shape: .circle,
+      ),
+      child: Container(
+        width: 64,
+        height: 64,
+        clipBehavior: Clip.antiAlias,
+        decoration: BoxDecoration(
+          color: Colors.amber,
+          shape: .circle,
+        ),
+        child: Image.asset(
+          'assets/bird.png',
+          fit: BoxFit.cover,
+        ),
+      ),
+    );
+  }
+
+  Widget _networkImageMarker() {
+    return Container(
+      margin: const EdgeInsets.all(2),
+      padding: EdgeInsets.all(8),
+      decoration: BoxDecoration(
+        color: Colors.amber,
+        shape: .circle,
+      ),
+      child: Container(
+        width: 64,
+        height: 64,
+        clipBehavior: Clip.antiAlias,
+        decoration: BoxDecoration(
+          color: Colors.amber,
+          shape: .circle,
+        ),
+        child: Image.network(
+          'https://docs.flutter.dev/assets/images/dash/Dash.png',
+          fit: BoxFit.cover,
+        ),
+      ),
+    );
+  }
+
+  Future<void> _cacheAssets() async {
+    await Future.wait([
+      _hybridMapMarker.cacheSvg(path: 'assets/user.svg'),
+      _hybridMapMarker.cacheImage(path: 'assets/bird.jpg'),
+      _hybridMapMarker.cacheImage(path: 'assets/bird.png'),
+      _hybridMapMarker.cacheNetworkImage(
+        path: 'https://docs.flutter.dev/assets/images/dash/Dash.png',
+      ),
+    ]);
+  }
+
   Future<Set<Marker>> _createMarkers() async {
-    await _hybridMapMarker.cacheSvg(path: 'assets/user.svg');
+    await _cacheAssets();
 
     final size = const Size(64, 64);
     final locationIcon = await _hybridMapMarker.createIcon(
@@ -80,6 +164,9 @@ class _MainAppState extends State<MainApp> {
       size: size,
     );
     final svgIcon = await _hybridMapMarker.createIcon(_svgMarker(), size: size);
+    final jpgIcon = await _hybridMapMarker.createIcon(_jpgImageMarker(), size: size);
+    final pngIcon = await _hybridMapMarker.createIcon(_pngImageMarker(), size: size);
+    final networkIcon = await _hybridMapMarker.createIcon(_networkImageMarker(), size: size);
 
     return {
       Marker(
@@ -91,6 +178,21 @@ class _MainAppState extends State<MainApp> {
         markerId: MarkerId('svg'),
         position: const LatLng(39.482232741106344, -0.3567805203800181),
         icon: svgIcon,
+      ),
+      Marker(
+        markerId: MarkerId('jpg'),
+        position: const LatLng(39.479932432434474, -0.3556799773645045),
+        icon: jpgIcon,
+      ),
+      Marker(
+        markerId: MarkerId('png'),
+        position: const LatLng(39.48056070220794, -0.35663411718664756),
+        icon: pngIcon,
+      ),
+      Marker(
+        markerId: MarkerId('network'),
+        position: const LatLng(39.481361447965654, -0.3557271053898057),
+        icon: networkIcon,
       ),
     };
   }